Added test for different variant and condition
Added an example to A - Spec Features\h_Loops.feature with a variant that is different than the loop condition. Also in the example to multiply by adding, changed the while test back to using < in order to show how the codition combines with the invariant to support the desired loop meaning.